feat: provide wrapping interface for logging service
The logging service now provides an interface that allows a library or
application to inject any concrete implementation of a logging service.
See the README for details.

Additionally, renamed LoggingeService to NewRelicLoggingService, to make
it more clear that it is a concrete implementation that relies on
NewRelic.

BREAKING CHANGE: The LoggingService was renamed to
NewRelicLoggingService, and the new expected usage of the library is to
use the wrapped functions as described in the README.

ARCH-687