feat: update logging of axios errors (#14)
Splits logging and processing of axios errors apart into two functions. Errors that do not contain a response object are still use logInfo instead of logError. Custom attributes can be supplied to the error and a custom messagePrefix will be prepended to the error message.

BREAKING CHANGE: logAPIErrorResponse(error, customAttrs) is renamed to logAxiosError(error, customAttrs). It uses processAxiosError which clients can also leverage to normalize axios error objects. logAxiosError will take a messagePrefix specified in customAttrs and prepend that to error messages.