fix: refactor ensurePublicOrAuthenticationAndCookies
- Protects user against an infinite redirect to login
- Refactors ensurePublicOrAuthenticationAndCookies with the
below breaking changes.

BREAKING CHANGE: The refactor has the following breaking changes:

- Renamed ensurePublicOrAuthenticationAndCookies to
ensureAuthenticatedUser.
- Removed the capability to handle public routes in
ensureAuthenticatedUser.  Just don't call it.
- Removed getAuthenticationState() function which was unreliable,
because it used the cookie which could have expired since it was tested.
Instead, use the new response sent to the ensureAuthenticatedUser
promise (see below).
- Removed the deprecated callback.  Just use the promise instead.

The promise is now resolved with an object of the form:
```
{
  authenticatedUser: {...},
  decodedAccessToken: {...},
}

In the above object:
- ``authenticatedUser`` is an object containing user account data that
was stored in the access token.
- ``decodedAccessToken`` is the raw version of the data used to create
``authenticatedUser``.

ARCH-948