feat: refactor csrf and jwt roken refreshing and tests
BREAKING CHANGE: frontend-logging is no longer a direct dependency of this package. As a result you must make sure it is configured prior to calling getAuthenticatedAPIClient. In all likelihood all users of this package we already doing this. This change removes a kind of doubling up on configuration of the logging service.

* feat: simplify access token refresh queuing

* test: update authInterface tests

* refactor: breakout access token logic from authInterface

* test: leverage axios mock for tests

* refactor: csrf token manager

* fix: naming of customattributes