fix: correctly passing URL through to auth during initialization
We weren’t passing the current URL through to ensureAuthenticatedUser during initialization, which meant that - if the user was logged out - we weren’t redirecting to the right page post-login.

Also removing FormattedHTMLMessage as an export - it was a mistake that it was available, as it’s not XSS safe.