Extend getConfig
to optionally lookup the value both in config now also check process.env .
An optional fallback/default can also be provided.

So the current use-case should still work, assuming you've exported the
ENV variable(s):

```js
const url = getConfig().URL;
const {
  URL,
} = getConfig();
```

This change also enables a key-based lookup that coalesces, looking for
the key in:
- config
- process.env

If still not _defined_ anywhere, it will fallback to a default value.

```js
const url = getConfig('URL');
const URL = getConfig('URL', 'localhost:18000');
```

[...]

In the future, this could be extended to check for the setting in a YAML file.