Improving shape of auth configure.
Still haven’t cut over to an interface, and am still exporting AxiosJwtAuthService directly from index.js.

The next step is going to be to start turning AxiosJwtAuthService into a class so that it can be instantiated.  That’ll be the hard part.

Once that’s done, we can cut index.js over to use interface.js (still not checked in).  Then the configure() function in interface.js can instantiate the AuthService class properly.