feat: Auth module interface creation (#91)
NOTE: This PR technically includes a backwards incompatible change to the method signature of the configure function in the frontend-platform/auth package. I've chosen NOT to mark this as a major version bump of the package because I can find no evidence in the project's dependents that any external projects are using this API. To do so, they would have had to forego using the initialize function and instead manually configured the auth service. If you're doing this and seeing errors that various types of configuration are undefined, please adjust your call to configure.

* Moving auth service out of index.

Next is to create an interface.js file.

* Improving shape of auth configure.

Still haven’t cut over to an interface, and am still exporting AxiosJwtAuthService directly from index.js.

The next step is going to be to start turning AxiosJwtAuthService into a class so that it can be instantiated.  That’ll be the hard part.

Once that’s done, we can cut index.js over to use interface.js (still not checked in).  Then the configure() function in interface.js can instantiate the AuthService class properly.

* First pass at creating AxiosJwtAuthService.

The service still acts as a singleton in some respects, but it can at least be instantiated.  interceptors, creating authenticated HTTP clients, and getting the JWT and CSRF tokens, still act as singletons/functions.  Next step is to incorporate those into the class.

* Folding addAuthenticationToHttpClient in as a static method.

* Converting getJwtToken into a class.

There’s now an AxiosJwtTokenService which does what getJwtToken.js used to do.  It’s instantiated by the auth service, and then used by both the JWT token provider interceptor and the auth service itself.

* All auth singletons gone - jk, one is left.

The hold-out was the loggingService reference, which is now passed through to the JWT token service and the interceptors, rather than imported directly into them from the auth service implementation itself.

AxiosJwtAuthService is now a proper class that can be instanced, except for getCsrfToken.

* Normalizing config variable names in auth

Now they match the values returned by getConfig().

* NOW the auth service is a proper set of classes.

I forgot about the CSRF token code.  Now it’s a class similar to AxiosJwtTokenService.

* Clean up and fixing broken docs.

The @implements and @memberof tags need to be attached to the class, not floating up at the top of the file.

* Updating README since auth is configurable now.