feat: Adding a MockAuthService, improving docs
The mock auth service uses axios-mock-adapter in its implementation to allow mocking of any http client calls the auth service might make on behalf of an application.

This PR also fleshes out the docs for the auth service to try to add some detail around the differences between the interface, AxiosJwtAuthService, and new MockAuthService implementations.