fix: MockAuthService returns normal axios instances
MockAuthService misused axios-mock-adapter’s MockAdapter class.  It is not a drop-in replacement for an axios instance, but a companion object that, when configured, will intercept requests to the axios instance it was instantiated with.  You cannot call .get() or .post() on the mock itself - you need to call those on the original instance.  Therefore MockAuthService needs to return the real deal, and instantiators are responsible for using axios-mock-adapter if they want to.

This means the class’s primary use is to mock out the authenticated user loading.