fix: allowing func or object as proptypes for httpClient
The analytics service needs an httpClient, and depending on whether we’re using an axios mock or a real axios instance, the type of that client is different.  There’s no good way around this, so this removes a warning that we otherwise would have gotten in the test suite or in development.  In either case, it was annoying.