fix: subtle bug in hasIdentifyBeenCalled variable
I noticed this in a test - we were accessing hasIdentifyBeenCalled as an instance variable, but it was declared as static.  This meant that if a test wanted to assert it was false, it failed because it was actually `undefined`.  It wasn’t a problem in prod because `undefined` is falsy, but I didn’t like the imprecision.