fix: Conditionally initialize Segment only if a SEGMENT_KEY is present. (#168)
* fix: allowing func or object as proptypes for httpClient

The analytics service needs an httpClient, and depending on whether we’re using an axios mock or a real axios instance, the type of that client is different.  There’s no good way around this, so this removes a warning that we otherwise would have gotten in the test suite or in development.  In either case, it was annoying.

* fix: subtle bug in hasIdentifyBeenCalled variable

I noticed this in a test - we were accessing hasIdentifyBeenCalled as an instance variable, but it was declared as static.  This meant that if a test wanted to assert it was false, it failed because it was actually `undefined`.  It wasn’t a problem in prod because `undefined` is falsy, but I didn’t like the imprecision.

* fix: only initialize Segment if a SEGMENT_KEY exists

This gates segment initialization on the presence of a valid SEGMENT_KEY from configuration.  A truthy SEGMENT_KEY is assumed to be valid.

If segment isn’t initialized, the SegmentAnalyticsService now short circuits any Segment analytics functions that might have been called.  Non-segment service methods will still work. (i.e., tracking log events)

Not having a SEGMENT_KEY should be a valid configuration, especially for Open edX operators who don’t use segment.  Given that micro-frontend consumers of the analytics interface _assume_ that they can call it with impunity and without error, the service now errs on the side of silently doing nothing, rather than complaining if SEGMENT_KEY isn’t present and a Segment-related function is called.