fix: Guard against a missing IGNORED_ERROR_REGEX environment variable.
The IGNORED_ERROR_REGEX variable is an optional parameter, so the config and
logging code should work fine without the variable being defined. As MFEs upgrade
to the newest version of frontend-platform, the new version should *not* require
the addition of IGNORED_ERROR_REGEX to function properly. This commit fixes the
previous code, which inadvertently required this variable to be defined.