fix: Properly handle an error object sent into logInfo().
At several places in certain MFEs, logInfo() is called with a caught error
object. But logInfo() has always assumed that only a string will be passed-in.
This situation causes a JSON object to be send to New Relic as the "message"
of an info log, which makes the info logs difficult to search.

This change adds support for either an info string -or- and error object being
sent to logInfo(). The proper "message" is logged either way.

TNL-8335