fix: identifyAnonymousUser promise not resolving
This is an update to a previous fix:
https://github.com/edx/frontend-platform/pull/208

If the Segment key is invalid or the user has blocked segment domain
using a plugin or extension; the promise does not get resolved and
app initialization fails resulting into a blank page.

Set timeout to resolve the promise if script is not loaded successfully.

VAN-689