fix: ensure that SCSS files are not treeshaken
By including *.scss in sideEffects, we guarantee that imports of our SCSS files aren't getting tree shaken.  This manifested as a problem in the example app, where the example/index.scss file wasn't showing up in the dev server.  The import was getting tree shook (?) because it wasn't importing any specific export.