fix: bumping frontend-build to v9 to avoid webpack version skew (#265)
The prior version of frontend-build contained a dependency (html-webpack-new-relic-plugin) which had a peer dependency of webpack v4.x.  This caused frontend-platform to install webpack 4 in its top-level node_modules/webpack directory, which caused `npm start` to explode, since the webpack config schema changed between 4 and 5, and we depend on features of webpack 5.