fix: removing assets from Github release
We don't use these assets - npm is the official release that's used in consumers.  The Github release is just for informational purposes.  This is an attempt to fix an issue where our releases on GitHub are failing to complete, either because we have too many assets, or because our asset bundle includes files with duplicate names at different levels of the dist directory's tree, resulting in a ReleaseAsset "already_exists" error from GitHub/octokit.