fix: removing unused i18n/scripts/Makefile and adding README
See new src/i18n/scripts/README.md file.  The Makefile in this directory was never used - I added it years ago to attempt to “import” it into the Makefiles in MFEs; that approach never worked and I forgot to clean it up.

This commit cleans it up and adds the README.md to explain how transifex-utils.js works.