feat: track Segment event to quantify users' preferred color scheme (#444)
This releases intends to begin collecting data on users' preference for light vs. dark color scheme as input and/or validation to the prioritization around supporting a true dark mode for the Open edX platform. That is, we'll be able to quantify what percentage of users would prefer to use a dark mode even though we don't offer one on Open edX.