feat: JavaScript file configuration of apps (#474)
* feat: JavaScript file configuration of apps

This adds the ability to configure an application via an env.config.js file rather than environment variables or .env files.  This mechanism is much more flexible and powerful than environment variables as described in the associated ADR.

Note that this is not in _use_ by any MFEs as of this PR being merged - this merely adds the capability.  Guidance will be forthcoming on how to use this mechanism, along with a DEPR issue for the deprecation of environment variable configuration later this year.  

It also improves documentation around how to configure applications, providing more detail on the various methods.

Finally, it allows the logging, analytics, and auth services to be configured via the configuration document now that we can supply an alternate implementation in an env.config.js file.  This allows configuration of these services without forking the MFE.

We have agreed we’d like to DEPR environment variable config, so I’m going to merge the ADR as “Accepted”