test: cutting example apps over to use env.config.js
Each app can have its own env.config.js because they’re independently run.  We no longer need any of the .env files here.  Since we’re no longer using .env as an input into the webpack process, we need to hardcode the port for the plugins app to 8081 so it doesn’t fall back to 8080 and conflict with the main example app.