feat: add `getPath` function for `PUBLIC_PATH`
This add function to get path from a string url from utils.
Also the function is implemented to the history module. In order to match with URLS as webpack allow.(CDN).
Test were developed for different shaped of string urls with path.
A basic test is also added for the base case in the history variable for the initialize
with base case of PUBLIC_PATH=`/`.