fix: make Tutor-MFE config compatible with React Router v6
As part of the React Router v5 -> v6 upgrade, history was no longer
being passed in. This broke the Tutor deployment method of MFEs, which
puts all the MFEs in a common server and uses different top-level dirs
to separate them.

The fix for this is to pass in an explicit basename to the Router.