feat: allow customization of external URLs
Create the getExternalLinkUrl function and make it available for
use in frontend apps. It checks for a `customExternalUrls` object
in the config to see if a custom URL has been provided for a given
URL.

Use getExternalLinkUrl in the example app for this package. One
link is overridden with a custom URL while the other is not.